+++ tests/build_tests/devonly/input.js
@@ -2,6 +2,6 @@
 
 import { setup } from "#dev/process";
 
-const { execBuild } = setup(import.meta.dirname);
+const { execBuildOrThrow } = setup(import.meta.dirname);
 
-await execBuild();
+await execBuildOrThrow();