+++ tests/build_tests/unicode/input.js
@@ -4,14 +4,14 @@ import * as fs from "node:fs/promises";
 import * as path from "node:path";
 import { setup } from "#dev/process";
 
-const { execBuild, execClean } = setup(import.meta.dirname);
+const { execBuildOrThrow, execClean } = setup(import.meta.dirname);
 
 if (process.platform === "win32") {
   console.log("Skipping test on Windows");
   process.exit(0);
 }
 
-await execBuild();
+await execBuildOrThrow();
 
 await fs.access(
   path.join(import.meta.dirname, "lib", "bs", "src", "📕annotation", "a.js"),